more refactoring of files, changed a few api routes
