+++ utils/misc/generateCode.js
